remove cos
