+++ deps/init/oceanbase.el8.x86_64.deps
@@ -30,7 +30,6 @@ devdeps-boost-1.74.0-22022110914.el8.x86_64.rpm
 devdeps-abseil-cpp-20211102.0-62024122014.el8.x86_64.rpm
 devdeps-s2geometry-0.10.0-122024122617.el8.x86_64.rpm
 devdeps-icu-69.1-72022112416.el8.x86_64.rpm
-devdeps-cos-c-sdk-5.0.21-202024080916.el8.x86_64.rpm
 devdeps-s3-cpp-sdk-1.11.156-102023122011.el8.x86_64.rpm
 devdeps-protobuf-c-1.5.1-100000052025032810.el8.x86_64.rpm
 devdeps-roaringbitmap-croaring-3.0.0-42024042816.el8.x86_64.rpm
