+++ deps/oblib/src/common/storage/ob_device_common.h
@@ -96,12 +96,11 @@ enum ObStorageType : uint8_t
 {
   OB_STORAGE_OSS = 0,
   OB_STORAGE_FILE = 1,
-  OB_STORAGE_COS = 2,
-  OB_STORAGE_LOCAL = 3,
-  OB_STORAGE_S3 = 4,
-  OB_STORAGE_LOCAL_CACHE = 5,
-  OB_STORAGE_HDFS = 6,
-  OB_STORAGE_AZBLOB = 7,
+  OB_STORAGE_LOCAL = 2,
+  OB_STORAGE_S3 = 3,
+  OB_STORAGE_LOCAL_CACHE = 4,
+  OB_STORAGE_HDFS = 5,
+  OB_STORAGE_AZBLOB = 6,
   OB_STORAGE_MAX_TYPE
 };
