+++ deps/oblib/src/common/storage/ob_io_device.h
@@ -460,7 +460,6 @@ public:
   OB_INLINE bool is_object_device() const
   {
     return (ObStorageType::OB_STORAGE_OSS == device_type_)
-           || (ObStorageType::OB_STORAGE_COS == device_type_)
            || (ObStorageType::OB_STORAGE_S3 == device_type_)
            || (ObStorageType::OB_STORAGE_FILE == device_type_)
            || (ObStorageType::OB_STORAGE_AZBLOB == device_type_)
