+++ deps/oblib/src/lib/restore/ob_storage.cpp
@@ -26,7 +26,7 @@ namespace oceanbase
 namespace common
 {
 
-const char *OB_STORAGE_TYPES_STR[] = {"OSS", "FILE", "COS", "LOCAL", "S3", "LOCAL_CACHE", "HDFS", "AZBLOB"};
+const char *OB_STORAGE_TYPES_STR[] = {"OSS", "FILE", "LOCAL", "S3", "LOCAL_CACHE", "HDFS", "AZBLOB"};
 
 void print_access_storage_log(
     const char *msg,
@@ -57,7 +57,6 @@ int validate_uri_type(const common::ObString &uri)
 {
   int ret = OB_SUCCESS;
   if (!uri.prefix_match(OB_OSS_PREFIX) &&
-      !uri.prefix_match(OB_COS_PREFIX) &&
       !uri.prefix_match(OB_S3_PREFIX) &&
       !uri.prefix_match(OB_FILE_PREFIX) &&
       !uri.prefix_match(OB_HDFS_PREFIX) &&
@@ -75,8 +74,6 @@ int get_storage_type_from_path(const common::ObString &uri, ObStorageType &type)
 
   if (uri.prefix_match(OB_OSS_PREFIX)) {
     type = OB_STORAGE_OSS;
-  } else if (uri.prefix_match(OB_COS_PREFIX)) {
-    type = OB_STORAGE_COS;
   } else if (uri.prefix_match(OB_S3_PREFIX)) {
     type = OB_STORAGE_S3;
   } else if (uri.prefix_match(OB_FILE_PREFIX)) {
@@ -105,7 +102,6 @@ const char *get_storage_type_str(const ObStorageType &type)
 bool is_storage_type_match(const common::ObString &uri, const ObStorageType &type)
 {
   return (OB_STORAGE_OSS == type && uri.prefix_match(OB_OSS_PREFIX))
-      || (OB_STORAGE_COS == type && uri.prefix_match(OB_COS_PREFIX))
       || (OB_STORAGE_S3 == type && uri.prefix_match(OB_S3_PREFIX))
       || (OB_STORAGE_FILE == type && uri.prefix_match(OB_FILE_PREFIX))
       || (OB_STORAGE_HDFS == type && uri.prefix_match(OB_HDFS_PREFIX))
@@ -505,7 +501,6 @@ ObExternalIOCounterGuard::~ObExternalIOCounterGuard()
 ObStorageUtil::ObStorageUtil()
   : file_util_(),
     oss_util_(),
-    cos_util_(),
     s3_util_(),
     hdfs_util_(),
     util_(NULL),
@@ -534,8 +529,6 @@ int ObStorageUtil::open(common::ObObjectStorageInfo *storage_info)
     util_ = &hdfs_util_;
   } else if (OB_STORAGE_OSS == device_type_) {
     util_ = &oss_util_;
-  } else if (OB_STORAGE_COS == device_type_) {
-    util_ = &cos_util_;
   } else if (OB_STORAGE_S3 == device_type_) {
     util_ = &s3_util_;
   } else {
@@ -1861,7 +1854,6 @@ ObStorageReader::ObStorageReader()
     reader_(NULL),
     file_reader_(),
     oss_reader_(),
-    cos_reader_(),
     s3_reader_(),
     start_ts_(0),
     has_meta_(false),
@@ -1913,8 +1905,6 @@ int ObStorageReader::open(const common::ObString &uri,
     reader_ = &hdfs_reader_;
   } else if (OB_STORAGE_OSS == type) {
     reader_ = &oss_reader_;
-  } else if (OB_STORAGE_COS == type) {
-    reader_ = &cos_reader_;
   } else if (OB_STORAGE_S3 == type) {
     reader_ = &s3_reader_;
   } else {
@@ -2017,7 +2007,6 @@ ObStorageAdaptiveReader::ObStorageAdaptiveReader()
       reader_(NULL),
       file_reader_(),
       oss_reader_(),
-      cos_reader_(),
       s3_reader_(),
       hdfs_reader_(),
       start_ts_(0),
@@ -2059,8 +2048,6 @@ static int alloc_reader(ObIAllocator &allocator, const ObStorageType &type, ObIS
     ret = alloc_reader_type<ObStorageHdfsReader>(allocator, reader);
   } else if (OB_STORAGE_OSS == type) {
     ret = alloc_reader_type<ObStorageOssReader>(allocator, reader);
-  } else if (OB_STORAGE_COS == type) {
-    ret = alloc_reader_type<ObStorageCosReader>(allocator, reader);
   } else if (OB_STORAGE_S3 == type) {
     ret = alloc_reader_type<ObStorageS3Reader>(allocator, reader);
   }
@@ -2108,8 +2095,6 @@ int ObStorageAdaptiveReader::open(const common::ObString &uri,
     reader_ = &hdfs_reader_;
   } else if (OB_STORAGE_OSS == type) {
     reader_ = &oss_reader_;
-  } else if (OB_STORAGE_COS == type) {
-    reader_ = &cos_reader_;
   } else if (OB_STORAGE_S3 == type) {
     reader_ = &s3_reader_;
   } else {
@@ -2283,7 +2268,6 @@ ObStorageWriter::ObStorageWriter()
   : writer_(NULL),
     file_writer_(),
     oss_writer_(),
-    cos_writer_(),
     s3_writer_(),
     start_ts_(0),
     storage_info_(nullptr)
@@ -2331,8 +2315,6 @@ int ObStorageWriter::open(const common::ObString &uri, common::ObObjectStorageIn
     writer_ = &file_writer_;
   } else if (OB_STORAGE_OSS == type) {
     writer_ = &oss_writer_;
-  } else if (OB_STORAGE_COS == type) {
-    writer_ = &cos_writer_;
   } else if (OB_STORAGE_S3 == type) {
     writer_ = &s3_writer_;
   } else {
@@ -2411,7 +2393,6 @@ ObStorageAppender::ObStorageAppender()
   : appender_(NULL),
     file_appender_(),
     oss_appender_(),
-    cos_appender_(),
     s3_appender_(),
     start_ts_(0),
     is_opened_(false),
@@ -2426,7 +2407,6 @@ ObStorageAppender::ObStorageAppender(StorageOpenMode mode)
   : appender_(NULL),
     file_appender_(mode),
     oss_appender_(),
-    cos_appender_(),
     s3_appender_(),
     start_ts_(0),
     is_opened_(false),
@@ -2477,12 +2457,9 @@ int ObStorageAppender::open(
   } else if (OB_STORAGE_FILE == type_) {
     appender_ = &file_appender_;
   } else if (OB_STORAGE_OSS == type_ || 
-             OB_STORAGE_COS == type_ || 
              OB_STORAGE_S3 == type_) {
     if (OB_STORAGE_OSS == type_) {
       appender_ = &oss_appender_;
-    } else if (OB_STORAGE_COS == type_) {
-      appender_ = &cos_appender_;
     } else if (OB_STORAGE_S3 == type_) {
       appender_ = &s3_appender_;
     }
@@ -2724,7 +2701,6 @@ int ObStorageAppender::seal_for_adaptive()
 ObStorageMultiPartWriter::ObStorageMultiPartWriter()
     : multipart_writer_(NULL),
       file_multipart_writer_(),
-      cos_multipart_writer_(),
       oss_multipart_writer_(),
       s3_multipart_writer_(),
       start_ts_(0),
@@ -2772,12 +2748,9 @@ int ObStorageMultiPartWriter::open(
   } else if (OB_STORAGE_FILE == type) {
     multipart_writer_ = &file_multipart_writer_;
   } else if (OB_STORAGE_OSS == type || 
-             OB_STORAGE_COS == type || 
              OB_STORAGE_S3 == type) {
     if (OB_STORAGE_OSS == type) {
       multipart_writer_ = &oss_multipart_writer_;
-    } else if (OB_STORAGE_COS == type) {
-      multipart_writer_ = &cos_multipart_writer_;
     } else if (OB_STORAGE_S3 == type) {
       multipart_writer_ = &s3_multipart_writer_;
     }
@@ -2953,7 +2926,6 @@ int ObStorageMultiPartWriter::close()
 ObStorageParallelMultiPartWriterBase::ObStorageParallelMultiPartWriterBase()
     : multipart_writer_(nullptr),
       file_multipart_writer_(),
-      cos_multipart_writer_(),
       oss_multipart_writer_(),
       s3_multipart_writer_(),
       start_ts_(0),
@@ -3004,8 +2976,6 @@ int ObStorageParallelMultiPartWriterBase::open(
     multipart_writer_ = &file_multipart_writer_;
   } else if (OB_STORAGE_OSS == type) {
     multipart_writer_ = &oss_multipart_writer_;
-  } else if (OB_STORAGE_COS == type) {
-    multipart_writer_ = &cos_multipart_writer_;
   } else if (OB_STORAGE_S3 == type) {
     multipart_writer_ = &s3_multipart_writer_;
   } else {
