+++ deps/oblib/src/lib/restore/ob_storage.h
@@ -19,7 +19,6 @@
 #include "ob_i_storage.h"
 #include "ob_storage_file.h"
 #include "ob_storage_oss_base.h"
-#include "ob_storage_cos_base.h"
 #include "ob_storage_s3_base.h"
 #include "hdfs/ob_storage_hdfs_jni_base.h"
 #include "common/storage/ob_io_device.h"
@@ -276,7 +275,7 @@ public:
       const ObIArray<ObString> &files_to_delete, ObIArray<int64_t> &failed_files_idx);
   int del_unmerged_parts(const common::ObString &uri);
 
-  // For one object, if given us the uri(no matter in oss, cos or s3), we can't tell the type of this object.
+  // For one object, if given us the uri(no matter in oss or s3), we can't tell the type of this object.
   // It may be a 'single、normal' object. Or it may be a 's3-appendable-object'(like a dir), containing several 
   // 'single、normal' objects.
   // So, this function is for checking the object meta, to get its meta info
@@ -359,7 +358,6 @@ private:
 
   ObStorageFileUtil file_util_;
   ObStorageOssUtil oss_util_;
-  ObStorageCosUtil cos_util_;
   ObStorageS3Util s3_util_;
   ObStorageHdfsJniUtil hdfs_util_;
   ObIStorageUtil* util_;
@@ -432,7 +430,6 @@ protected:
   ObIStorageReader *reader_;
   ObStorageFileReader file_reader_;
   ObStorageOssReader oss_reader_;
-  ObStorageCosReader cos_reader_;
   ObStorageS3Reader s3_reader_;
   ObStorageHdfsReader hdfs_reader_;
   int64_t start_ts_;
@@ -463,7 +460,6 @@ private:
   ObIStorageReader *reader_;
   ObStorageFileReader file_reader_;
   ObStorageOssReader oss_reader_;
-  ObStorageCosReader cos_reader_;
   ObStorageS3Reader s3_reader_;
   ObStorageHdfsReader hdfs_reader_;
   int64_t start_ts_;
@@ -484,7 +480,6 @@ protected:
   ObIStorageWriter *writer_;
   ObStorageFileSingleWriter file_writer_;
   ObStorageOssWriter oss_writer_;
-  ObStorageCosWriter cos_writer_;
   ObStorageS3Writer s3_writer_;
   int64_t start_ts_;
   char uri_[OB_MAX_URI_LENGTH];
@@ -519,7 +514,6 @@ private:
   ObIStorageWriter *appender_;
   ObStorageFileAppender file_appender_;
   ObStorageOssAppendWriter oss_appender_;
-  ObStorageCosAppendWriter cos_appender_;
   ObStorageS3AppendWriter s3_appender_;
   int64_t start_ts_;
   bool is_opened_;
@@ -551,7 +545,6 @@ public:
 protected:
   ObIStorageMultiPartWriter *multipart_writer_;
   ObStorageFileMultiPartWriter file_multipart_writer_;
-  ObStorageCosMultiPartWriter cos_multipart_writer_;
   ObStorageOssMultiPartWriter oss_multipart_writer_;
   ObStorageS3MultiPartWriter s3_multipart_writer_;
   int64_t start_ts_;
@@ -574,7 +567,6 @@ public:
 protected:
   ObIStorageParallelMultipartWriter *multipart_writer_;
   ObStorageParallelFileMultiPartWriter file_multipart_writer_;
-  ObStorageParallelCosMultiPartWriter cos_multipart_writer_;
   ObStorageParallelOssMultiPartWriter oss_multipart_writer_;
   ObStorageParallelS3MultiPartWriter s3_multipart_writer_;
   int64_t start_ts_;
