+++ deps/oblib/unittest/lib/CMakeLists.txt
@@ -91,7 +91,6 @@ oblib_addtest(resource/test_resource_mgr.cpp)
 oblib_addtest(resource/test_affinity_ctrl.cpp)
 #oblib_addtest(restore/test_storage_file.cpp)
 #oblib_addtest(restore/test_storage_oss.cpp)
-#oblib_addtest(restore/test_storage_cos.cpp)
 #oblib_addtest(restore/test_storage_s3.cpp)
 # oblib_addtest(restore/test_object_storage.cpp)
 # oblib_addtest(restore/test_common_storage.cpp)
