+++ unittest/share/backup/CMakeLists.txt
@@ -1,6 +1,5 @@
 storage_unittest(test_backup_path)
 storage_unittest(test_backup_struct)
 storage_unittest(test_log_archive_backup_info_mgr)
-storage_unittest(test_backup_access_cos)
 storage_unittest(test_backup_access_s3)
 storage_unittest(test_archive_checkpoint_mgr)
