--- unittest/share/backup/test_backup_access_cos.h
@@ -1,27 +1,0 @@
-/*
- * Copyright (c) 2025 OceanBase.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#ifndef UNITTEST_BACKUP_ACCESS_COS_H_
-#define UNITTEST_BACKUP_ACCESS_COS_H_
-
-const bool enable_test = false;
-const char *bucket = "cos://xxx";
-const char *endpoint = "xxx";
-const char *secretid = "xxx";
-const char *secretkey = "xxx";
-const char *appid = "xxx";
-
-#endif
